# contracts/openapi.v1.yaml
openapi: 3.0.3
info:
  title: PM Agent API
  version: 1.0.0
  description: |
    Versioned contract between the Next.js frontend and FastAPI backend.
    Breaking changes require /v2. Authenticated routes send `Authorization: Bearer <access_token>`
    and optional `X-Workspace-Id`.
servers:
  - url: http://localhost:8000/v1
...
paths:
  /auth/signup:
    post:
      tags: [auth]
      summary: Create account, workspace, and default settings
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SignUpRequest" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokenResponse" }
  /auth/signin:
    post:
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SignInRequest" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokenResponse" }
  /auth/google:
    post:
      tags: [auth]
      summary: Sign in or create an account with a Google (Gmail) authorization code
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GoogleAuthRequest" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokenResponse" }
  /auth/refresh:
    post:
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties: { refresh_token: { type: string } }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokenResponse" }
  /auth/signout:
    post:
      tags: [auth]
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Signed out } }
  /auth/forgot-password:
    post:
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties: { email: { type: string, format: email } }
      responses: { "200": { description: Always ok } }
  /auth/reset-password:
    post:
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token: { type: string }
                password: { type: string, minLength: 8 }
      responses: { "200": { description: Password updated } }
  /auth/verify-email:
    post:
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties: { token: { type: string } }
      responses: { "200": { description: Verified } }
  /auth/me:
    get:
      tags: [auth]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserResponse" }
...
  /automation:
    get:
      tags: [product]
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Automation settings } }
    patch:
      tags: [product]
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Updated } }
  /automation/preset:
    post:
      tags: [product]
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Preset applied } }
  /automation/rules:
    post:
      tags: [product]
      security: [{ bearerAuth: [] }]
      responses: { "201": { description: Rule added } }
  /automation/rules/preview:
    post:
      tags: [product]
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Matching tickets } }
  /automation/rules/{rule_id}:
    delete:
      tags: [product]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: rule_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Removed } }
  /automation/rules/{rule_id}/toggle:
    post:
      tags: [product]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: rule_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Toggled } }
...
  /dispatch:
    get:
      tags: [product]
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Dispatch records } }
    post:
      tags: [product]
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Dispatched } }
  /dispatch/config:
    get:
      tags: [product]
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Config } }
    patch:
      tags: [product]
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Updated } }
  /chat/sessions:
    get:
      tags: [product]
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Sessions } }
    post:
      tags: [product]
      security: [{ bearerAuth: [] }]
      responses: { "201": { description: Created } }
  /chat/sessions/{session_id}:
    patch:
      tags: [product]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: session_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Renamed } }
    delete:
      tags: [product]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: session_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Deleted } }
  /chat/sessions/{session_id}/messages:
    get:
      tags: [product]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: session_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Messages } }
    post:
      tags: [product]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: session_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: User + PM messages } }
  /chat/sessions/{session_id}/messages/{message_id}/send-to-dev:
    post:
      tags: [product]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: session_id, in: path, required: true, schema: { type: string } }
        - { name: message_id, in: path, required: true, schema: { type: string } }
      responses: { "200": { description: Draft ticket created } }
  /insights/overview:
    get:
      tags: [product]
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: KPIs } }
  /insights/series:
    get:
      tags: [product]
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Time series } }
  /insights/ai-performance:
    get:
      tags: [product]
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Override/edit rates } }
  /alerts:
    get:
      tags: [product]
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: Alerts } }
  /alerts/{alert_id}/dismiss:
    post:
      tags: [product]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: alert_id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: Dismissed } }
  /audit:
    get:
      tags: [product]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: q, in: query, schema: { type: string } }]
      responses: { "200": { description: Audit log } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SignUpRequest:
      type: object
      required: [name, email, password, company]
      properties:
        name: { type: string }
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        company: { type: string }
    SignInRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    GoogleAuthRequest:
      type: object
      required: [code]
      properties:
        code: { type: string, description: Google OAuth authorization code from the Gmail redirect callback }
        company: { type: string, description: Optional workspace name used only when creating a new account }
...
    UserResponse:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        email_verified: { type: boolean }
        default_workspace_id: { type: string, nullable: true }
    AuthTokenResponse:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string }
        user: { $ref: "#/components/schemas/UserResponse" }
...